Updated the sentiment-analysis workflow as per new plasma version
