+++ sentiment-analysis/components/data_fetching/requirements.txt
