+++ sentiment-analysis/components/model_evaluation/component.py
@@ -1,0 +1,20 @@
+import tensorflow as tf
+import logging
+
+logger = logging.getLogger('model_evaluator')
+
+def evaluate(parameters):
+    logger.info('evaluating model loss')
+    X_test = parameters["X_test"]
+    y_test = parameters["y_test"]
+    batch_size = parameters["batch_size"]
+    model = parameters["model"]
+    result = model.evaluate(X_test, y_test, batch_size, verbose=1)
+    logger.info('model loss : '+str(result))
+    return True
+
+
+def main(args):
+    if args['component'] == 'model_evaluation':
+        output = evaluate(args['parameters'])
+        return output
