+++ sentiment-analysis/components/model_training/requirements.txt
