+++ sentiment-analysis/workflows/sentiment-analysis.yml
@@ -2,24 +2,24 @@ name: Sentiment Analysis Demo
 description: Sentiment Analysis Demo Pipeline
 version: 0.1
 workflow:
-  data_fetcher:
+  data_fetching:
     top_words: 1000
-  data_preprocessor:
+  data_preprocessing:
     max_review_length: 80
     X_train: X_train
     X_test: X_test
-  model_trainer:
+  model_training:
     X_train: X_train
     X_test: X_test
     y_train: y_train
     y_test: y_test
-    batch_size: 12
+    batch_size: 32
     epochs: 1
     units: 64
     top_words: 1000
     max_review_length: 80
     num_classes: 2
-  model_evaluator:
+  model_evaluation:
     model: model
     X_test: X_train
     y_test: y_test
