+++ social-distancing/components/inferencing/.gitignore
@@ -1,0 +1,1 @@
+/models/yolov3.weights
